+++ src/Rtp/RtpSplitter.h
@@ -31,8 +31,8 @@ protected:
 
 protected:
     int64_t onRecvHeader(const char *data, uint64_t len) override;
-    const char *onSearchPacketTail(const char *data, int len) override;
-    const char *onSearchPacketTail_l(const char *data, int len);
+    const char *onSearchPacketTail(const char *data, uint64_t len) override;
+    const char *onSearchPacketTail_l(const char *data, uint64_t len);
 
 private:
     int _offset = 0;
