+++ src/Rtsp/RtspSplitter.h
@@ -47,8 +47,8 @@ protected:
      */
     virtual int64_t getContentLength(Parser &parser);
 protected:
-    const char *onSearchPacketTail(const char *data,int len) override ;
-    const char *onSearchPacketTail_l(const char *data,int len) ;
+    const char *onSearchPacketTail(const char *data,uint64_t len) override ;
+    const char *onSearchPacketTail_l(const char *data,uint64_t len) ;
     int64_t onRecvHeader(const char *data,uint64_t len) override;
     void onRecvContent(const char *data,uint64_t len) override;
 private:
